pyproject.toml
[project]
name = "pako-tts-engines"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "TTS>=0.22",
    "torch>=2.0",
    "torchaudio>=2.0",
    "lingua-language-detector>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "httpx>=0.27",
    "mypy>=1.8",
    "ruff>=0.3",
    "types-PyYAML>=6.0",
]

[project.scripts]
pako-tts-engines = "app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B", "C4", "SIM"]
ignore = ["ANN401"]  # Allow Any in *args/**kwargs

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
    "types-pyyaml>=6.0.12.20250915",
]
